:refactor: Change folder name to config instead of core and move babel dependencies to Dependencies
