+++ package.json
@@ -15,15 +15,13 @@
     "type": "git",
     "url": "git+https://github.com/pghiraldelli/aprendareact-backend.git"
   },
-  "author": "",
+  "author": "Patricia Ghiraldelli",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/pghiraldelli/aprendareact-backend/issues"
   },
   "homepage": "https://github.com/pghiraldelli/aprendareact-backend#readme",
   "devDependencies": {
-    "@babel/core": "^7.7.2",
-    "@babel/preset-env": "^7.7.1",
     "eslint": "^6.6.0",
     "eslint-plugin-jest": "^23.0.4",
     "jest": "^24.9.0",
@@ -32,6 +30,8 @@
     "supertest": "^4.0.2"
   },
   "dependencies": {
+    "@babel/core": "^7.7.2",
+    "@babel/preset-env": "^7.7.1",
     "@babel/register": "^7.7.0",
     "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.4",
