+++ src/config/server.js
@@ -3,7 +3,7 @@ import cookieSession from 'cookie-session'
 import bodyParser from 'body-parser'
 import cookieParser from 'cookie-parser'
 import cors from 'cors'
-import passport from '../core/passportConfig'
+import passport from './passport'
 
 const oneDayInMiliseconds = 24 * 60 * 60 * 100
 const cookieKey = process.env.COOKIE_KEY
